Cleanup: Move useScrollable from Scrollable.js to useScrollable.js and fix indentation
Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)
